modify

+++ map.c
@@ -2,16 +2,16 @@
 
 static float load_factor = 0.75;
 
-static int capacity = 8192;
+static unsigned long capacity = 8192;
 
 static map_element ** elements;
 
-static int size;
+static uint32_t size;
 
-static int key_len;
+static uint8_t key_len;
 
-static int map_hash(short *key) {
-    int h = 0;
+static uint32_t map_hash(uint16_t *key) {
+    unsigned int h = 0;
     for(int i = 0; i < key_len; i++) {
         h = 31 * h + key[i];
     }
@@ -19,9 +19,9 @@ static int map_hash(short *key) {
     return h & (capacity - 1);
 }
 
-static void _destroy(map_element ** elements, int capacity) {
+static void _destroy(map_element ** elements, uint32_t capacity) {
     map_element *p, *q;
-    for(int i = 0; i < capacity; i++) {
+    for(uint32_t i = 0; i < capacity; i++) {
         p = elements[i];
         while(p != NULL) {
             q = p->next;
@@ -36,13 +36,13 @@ static void _destroy(map_element ** elements, int capacity) {
 }
 
 static void extend(){
-    int last_capacity = capacity;
+    uint32_t last_capacity = capacity;
     capacity = capacity << 1;
     map_element ** last_elements = elements;
     size = 0;
     map_init(key_len);
     map_element *p;
-    for(int i = 0; i < last_capacity; i++) {
+    for(uint32_t i = 0; i < last_capacity; i++) {
         p = last_elements[i];
         while(p != NULL) {
             map_put(p->key, p->entity);
@@ -52,7 +52,7 @@ static void extend(){
     _destroy(last_elements, last_capacity);
 }
 
-void map_init(int len) {
+void map_init(uint8_t len) {
     key_len = len;
     elements = malloc(sizeof(map_element) * capacity);
     if (elements == NULL) {
@@ -64,9 +64,9 @@ void map_init(int len) {
     }
 }
 
-void map_put(short *k, void * entity) {
-    int len = sizeof(short) * FIRST_LEN;
-    short * key = malloc(len);
+void map_put(uint16_t *k, void * entity) {
+    int len = sizeof(uint16_t) * FIRST_LEN;
+    uint16_t * key = malloc(len);
     MEM_CHECK(key);
     memcpy(key, k, len);
     int index = map_hash(key);
@@ -105,7 +105,7 @@ void map_put(short *k, void * entity) {
     }
 }
 
-void map_for_each(void(*p_func)(short *,void *)) {
+void map_for_each(void(*p_func)(uint16_t *,void *)) {
     map_element * p;
     for(int i = 0; i < capacity; i++) {
         p = elements[i];
@@ -116,10 +116,10 @@ void map_for_each(void(*p_func)(short *,void *)) {
     }
 }
 
-void * map_get(short *key) {
+void * map_get(uint16_t *key) {
     map_element *element = elements[map_hash(key)];
     while(element != NULL) {
-        if (memcmp(element->key, key, key_len * sizeof(short)) == 0) {
+        if (memcmp(element->key, key, key_len * sizeof(uint16_t)) == 0) {
             return element->entity;
         }
         element = element->next;
